requirements.txt
# ── Data & DB ─────────────────────────────────────────────────────────────────
yfinance>=0.2.40
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0

# ── Econometrics ──────────────────────────────────────────────────────────────
arch>=6.3.0
statsmodels>=0.14.0
scipy>=1.12.0

# ── Simulation ────────────────────────────────────────────────────────────────
pyspark>=3.5.0

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit>=1.35.0
plotly>=5.22.0
